LPS-65920 move
